# services/auth-service/src/main/openapi/swagger.yaml
...
openapi: 3.0.1
info:
  title: Authentication and user API
  version: 1.0.0
  description: Authentication, user profiles, seller onboarding, role administration,
    and public signing-key discovery.
  contact:
    name: Ecommerce Platform maintainers
servers:
- url: http://localhost:8081
  description: Local auth-service direct access
paths:
  "/api/user/profile":
    get:
      tags:
      - User profile
      operationId: getProfile
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApiResponseUserProfileDto"
        '401':
          description: "Bearer-token filter rejection: revoked, non-user, or invalid-identity tokens return 401; the body may be empty, and revoked tokens may return plain text. MVC-level authentication failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
            text/plain:
              schema:
                type: string
        '403':
          description: "Security denial: anonymous requests receive 403 with an empty body when OAuth2 login is disabled, and callers without permission receive 403. With OAuth2 login enabled and a provider configured, anonymous access may instead redirect to sign-in (302). MVC-level authorization failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '302':
          description: Conditional sign-in redirect when OAuth2 login is enabled and a provider is configured; otherwise anonymous protected requests receive 403.
          headers:
            Location:
              schema:
                type: string
                format: uri
        '503':
          description: "When a Bearer token is present and the revocation blacklist check fails, JwtAuthFilter rejects the request with an empty response."
      summary: Get my profile
      description: Returns the authenticated user’s profile.
      security:
      - bearerAuth: []
    put:
      tags:
      - User profile
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UserUpdateRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApiResponseUserProfileDto"
        '400':
          description: Request validation failed.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: "Bearer-token filter rejection: revoked, non-user, or invalid-identity tokens return 401; the body may be empty, and revoked tokens may return plain text. MVC-level authentication failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
            text/plain:
              schema:
                type: string
        '403':
          description: "Security denial: anonymous requests receive 403 with an empty body when OAuth2 login is disabled, and callers without permission receive 403. With OAuth2 login enabled and a provider configured, anonymous access may instead redirect to sign-in (302). MVC-level authorization failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '302':
          description: Conditional sign-in redirect when OAuth2 login is enabled and a provider is configured; otherwise anonymous protected requests receive 403.
          headers:
            Location:
              schema:
                type: string
                format: uri
        '503':
          description: "When a Bearer token is present and the revocation blacklist check fails, JwtAuthFilter rejects the request with an empty response."
      summary: Update my profile
      description: Updates the authenticated user’s editable profile fields.
      security:
      - bearerAuth: []
    servers:
    - url: http://localhost:8081
      description: Local auth-service direct access
    - url: http://localhost:8080
      description: Local API gateway
...
  "/api/user/seller-application":
    get:
      tags:
      - User profile
      operationId: getMine
      responses:
        '200':
          description: Current user seller application.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApiResponseSellerApplicationResponse"
        '204':
          description: No seller application exists.
        '401':
          description: "Bearer-token filter rejection: revoked, non-user, or invalid-identity tokens return 401; the body may be empty, and revoked tokens may return plain text. MVC-level authentication failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
            text/plain:
              schema:
                type: string
        '403':
          description: "Security denial: anonymous requests receive 403 with an empty body when OAuth2 login is disabled, and callers without permission receive 403. With OAuth2 login enabled and a provider configured, anonymous access may instead redirect to sign-in (302). MVC-level authorization failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '302':
          description: Conditional sign-in redirect when OAuth2 login is enabled and a provider is configured; otherwise anonymous protected requests receive 403.
          headers:
            Location:
              schema:
                type: string
                format: uri
        '503':
          description: "When a Bearer token is present and the revocation blacklist check fails, JwtAuthFilter rejects the request with an empty response."
      summary: Get my seller application
      description: Returns the authenticated user’s seller application, if one exists.
      security:
      - bearerAuth: []
    post:
      tags:
      - User profile
      operationId: apply
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SellerApplicationRequest"
        required: true
      responses:
        '201':
          description: Seller application submitted.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApiResponseSellerApplicationResponse"
        '400':
          description: Request validation failed.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: "Bearer-token filter rejection: revoked, non-user, or invalid-identity tokens return 401; the body may be empty, and revoked tokens may return plain text. MVC-level authentication failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
            text/plain:
              schema:
                type: string
        '403':
          description: "Security denial: anonymous requests receive 403 with an empty body when OAuth2 login is disabled, and callers without permission receive 403. With OAuth2 login enabled and a provider configured, anonymous access may instead redirect to sign-in (302). MVC-level authorization failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '302':
          description: Conditional sign-in redirect when OAuth2 login is enabled and a provider is configured; otherwise anonymous protected requests receive 403.
          headers:
            Location:
              schema:
                type: string
                format: uri
        '503':
          description: "When a Bearer token is present and the revocation blacklist check fails, JwtAuthFilter rejects the request with an empty response."
      summary: Submit seller application
      description: Submits a seller application for the authenticated user.
      security:
      - bearerAuth: []
    servers:
    - url: http://localhost:8081
      description: Local auth-service direct access
    - url: http://localhost:8080
      description: Local API gateway
...
components:
  schemas:
    AddressDto:
      required:
      - city
      - country
      - fullName
      - phone
      - state
      - street
      - zipCode
      type: object
      properties:
        fullName:
          maxLength: 120
          type: string
          x-not-null-message: Full name is required
          x-field-extra-annotation: '@jakarta.validation.constraints.NotBlank(message = "Full name is required")'
          x-size-message: Full name must not exceed 120 characters
          x-jackson-json-include-policy: ALWAYS
        phone:
          pattern: "^\\+?[0-9 ()-]{6,20}$"
          type: string
          x-not-null-message: Phone number is required
          x-field-extra-annotation: '@jakarta.validation.constraints.NotBlank(message = "Phone number is required")'
          x-pattern-message: Invalid phone number
          x-jackson-json-include-policy: ALWAYS
        street:
          maxLength: 200
          type: string
          x-not-null-message: Street address is required
          x-field-extra-annotation: '@jakarta.validation.constraints.NotBlank(message = "Street address is required")'
          x-size-message: Street address must not exceed 200 characters
          x-jackson-json-include-policy: ALWAYS
        city:
          maxLength: 80
          type: string
          x-not-null-message: City is required
          x-field-extra-annotation: '@jakarta.validation.constraints.NotBlank(message = "City is required")'
          x-size-message: City must not exceed 80 characters
          x-jackson-json-include-policy: ALWAYS
        state:
          maxLength: 80
          type: string
          x-not-null-message: State/Province is required
          x-field-extra-annotation: '@jakarta.validation.constraints.NotBlank(message = "State/Province is required")'
          x-size-message: State must not exceed 80 characters
          x-jackson-json-include-policy: ALWAYS
        zipCode:
          maxLength: 20
          type: string
          x-not-null-message: Zip/Postal code is required
          x-field-extra-annotation: '@jakarta.validation.constraints.NotBlank(message = "Zip/Postal code is required")'
          x-size-message: Zip code must not exceed 20 characters
          x-jackson-json-include-policy: ALWAYS
        country:
          maxLength: 80
          type: string
          x-not-null-message: Country is required
          x-field-extra-annotation: '@jakarta.validation.constraints.NotBlank(message = "Country is required")'
          x-size-message: Country must not exceed 80 characters
          x-jackson-json-include-policy: ALWAYS
    UserUpdateRequest:
      type: object
      properties:
        displayName:
          maxLength: 80
          minLength: 2
          type: string
        imageUrl:
          type: string
        phone:
          pattern: "^\\+?[0-9 ()-]{6,20}$"
          type: string
          x-pattern-message: Invalid phone number
        shippingAddress:
          "$ref": "#/components/schemas/AddressDto"
        billingAddress:
          "$ref": "#/components/schemas/AddressDto"
    ApiResponseUserProfileDto:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          nullable: true
          allOf:
          - "$ref": "#/components/schemas/UserProfileDto"
          x-jackson-json-include-policy: ALWAYS
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
      - status
      - message
      - data
      - timestamp
    UserProfileDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          x-jackson-json-include-policy: ALWAYS
        email:
          type: string
          x-jackson-json-include-policy: ALWAYS
        displayName:
          type: string
          x-jackson-json-include-policy: ALWAYS
        imageUrl:
          type: string
          x-jackson-json-include-policy: ALWAYS
        phone:
          type: string
          x-jackson-json-include-policy: ALWAYS
        active:
          type: boolean
          x-jackson-json-include-policy: ALWAYS
        createdAt:
          type: string
          format: local-date-time
          x-jackson-json-include-policy: ALWAYS
        roles:
          uniqueItems: true
          type: array
          items:
            type: string
          x-jackson-json-include-policy: ALWAYS
        permissions:
          uniqueItems: true
          type: array
          items:
            type: string
          x-jackson-json-include-policy: ALWAYS
        shippingAddress:
          nullable: true
          allOf:
          - "$ref": "#/components/schemas/AddressDto"
          x-jackson-json-include-policy: ALWAYS
        billingAddress:
          nullable: true
          allOf:
          - "$ref": "#/components/schemas/AddressDto"
          x-jackson-json-include-policy: ALWAYS
        hasPassword:
          type: boolean
          x-jackson-json-include-policy: ALWAYS
...
    ApiResponseSellerApplicationResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          nullable: true
          allOf:
          - "$ref": "#/components/schemas/SellerApplicationResponse"
          x-jackson-json-include-policy: ALWAYS
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
      - status
      - message
      - data
      - timestamp
    SellerApplicationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          x-jackson-json-include-policy: ALWAYS
        userId:
          type: string
          format: uuid
          x-jackson-json-include-policy: ALWAYS
        userEmail:
          type: string
          x-jackson-json-include-policy: ALWAYS
        userDisplayName:
          type: string
          x-jackson-json-include-policy: ALWAYS
        status:
          type: string
          enum:
          - PENDING
          - APPROVED
          - REJECTED
          x-jackson-json-include-policy: ALWAYS
        businessName:
          type: string
          x-jackson-json-include-policy: ALWAYS
        gstin:
          type: string
          x-jackson-json-include-policy: ALWAYS
        contactPhone:
          type: string
          x-jackson-json-include-policy: ALWAYS
        pickupAddress:
          nullable: true
          allOf:
          - "$ref": "#/components/schemas/AddressDto"
          x-jackson-json-include-policy: ALWAYS
        bankAccountLast4:
          type: string
          x-jackson-json-include-policy: ALWAYS
        notes:
          type: string
          x-jackson-json-include-policy: ALWAYS
        rejectionReason:
          type: string
          x-jackson-json-include-policy: ALWAYS
        submittedAt:
          type: string
          format: local-date-time
          x-jackson-json-include-policy: ALWAYS
        reviewedAt:
          type: string
          format: local-date-time
          x-jackson-json-include-policy: ALWAYS
        reviewedBy:
          type: string
          format: uuid
          x-jackson-json-include-policy: ALWAYS
...
    SellerApplicationRequest:
      required:
      - businessName
      - contactPhone
      - pickupAddress
      type: object
      properties:
        businessName:
          maxLength: 180
          type: string
          x-not-null-message: must not be blank
          x-field-extra-annotation: '@jakarta.validation.constraints.NotBlank(message = "must not be blank")'
        gstin:
          pattern: "^[0-9A-Z]{15}$|^$"
          type: string
          x-pattern-message: GSTIN must be 15 alphanumeric characters
        contactPhone:
          pattern: "^\\+?[0-9 ()-]{6,20}$"
          type: string
          x-not-null-message: must not be blank
          x-field-extra-annotation: '@jakarta.validation.constraints.NotBlank(message = "must not be blank")'
          x-pattern-message: Invalid phone number
        pickupAddress:
          "$ref": "#/components/schemas/AddressDto"
        bankAccountLast4:
          pattern: "^[0-9]{4}$|^$"
          type: string
          x-pattern-message: Last-4 of bank account must be 4 digits
        notes:
          maxLength: 4000
          minLength: 0
          type: string
...
    ErrorResponse:
      type: object
      description: Standard error body returned by the shared exception handler.
      properties:
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        code:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...